A Traffic Light
Using JavaScript for interactivity, created a dynamic traffic light simulation that flashes red and green lights in response to user input. In order to simulate a traffic intersection, the project will display an elementary traffic light interface with buttons to turn on the red and green lights. The objective is to show off one's expertise in front-end development, namely in managing user interactions and using JavaScript to manipulate the DOM.